feat: build and test alpine binaries in system test

+++ system-test/system_test.sh
@@ -20,6 +20,8 @@ BASE_DIR=$(pwd)
 RUN_ONLY_V8_CANARY_TEST="${RUN_ONLY_V8_CANARY_TEST:-false}"
 echo "$RUN_ONLY_V8_CANARY_TEST"
 
+RUN_ON_ALPINE="${RUN_ON_ALPINE:-false}"
+
 # Install nvm.
 retry curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.34.0/install.sh | bash &>/dev/null
 export NVM_DIR="$HOME/.nvm" &>/dev/null
@@ -32,7 +34,17 @@ cp -R "system-test" "$TESTDIR"
 # Run test.
 cd "$TESTDIR"
 retry go get -t -d .
-go test -v -timeout=10m -run TestAgentIntegration -pprof_nodejs_path="$BASE_DIR" -run_only_v8_canary_test="$RUN_ONLY_V8_CANARY_TEST" -binary_host="$BINARY_HOST"
 
-# Remove directory where test was run.
-rm -r $TESTDIR
+if [ "$RUN_ON_ALPINE" == "true" ]; then
+  docker build -t test-image "alpine-docker"
+  docker run -v /var/run/docker.sock:/var/run/docker.sock \
+      -e GOPATH="$GOPATH" \
+      -v "$BASE_DIR":"$BASE_DIR" \
+      -v "$GOPATH":"$GOPATH" test-image \
+      go test "${BASE_DIR}/run-system-test/system_test.go" -v -timeout=10m -run TestAgentIntegration -pprof_nodejs_path="$BASE_DIR" -run_only_v8_canary_test="$RUN_ONLY_V8_CANARY_TEST" -binary_host="$BINARY_HOST"
+else
+  go test -v -timeout=10m -run TestAgentIntegration -pprof_nodejs_path="$BASE_DIR" -run_only_v8_canary_test="$RUN_ONLY_V8_CANARY_TEST" -binary_host="$BINARY_HOST"
+
+  # Remove directory where test was run.
+  rm -r $TESTDIR
+fi
